Added functionality to send email once feedback is submitted
Created the templates/mail directory for mail templates.  Created the new_feedback.html and .txt files which will be used when sending email.  Modified main.py to use flask mail.  Configured the email settings using environment variables.  Generated a secret key and moved that to environment variables.  Created the send_email function to simplify sending mail.  Modified the feedback view to grab the data from both the section and feedback areas of the form and send an email containing the data to the configured address.  Updated the requirements.txt since new libraries are being used.